% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_heatmap.R
\name{alignment_heatmap}
\alias{alignment_heatmap}
\title{Heatmap of aligned sequences}
\usage{
alignment_heatmap(files, nReads, AAthreshold = 0.9)
}
\description{
A heatmap of aligned sequences
}
\examples{
aa <- c('YNGTNQ', 'YGNMAKA')
alignment_heatmap(dna)
}
\keyword{AA,}
\keyword{heatmap}
