% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_tsne.R
\name{alignment_tsne}
\alias{alignment_tsne}
\title{Plot tsne from DNA/AA alignment}
\usage{
alignment_tsne(
  alignment,
  labeling = NULL,
  distance_metric = "hamming",
  perplexity = 30,
  num_threads = 12
)
}
\description{
This is a wrapper that uses a vector of same length strings to calculate a distance matrix which is used to calculate a dimensionality reduction via tSNE. This is then plotted with ggplot
}
\keyword{DNA,}
\keyword{complement}
\keyword{reverse,}
