% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignments2long.R
\name{alignments2long}
\alias{alignments2long}
\title{Convert alignment vector to long table}
\usage{
alignments2long(input, seqrange = NULL, get_frequencies = F)
}
\description{
This function takes a named vector of alignment sequences and converts it to a long data.frame with one letter per row. Columns indicate sequence ID and position in sequence. With get_frequencies the output can be condensed to frequency occurances of the letters per position.
}
\examples{
alns <- read_alignments(c('alnA.fa','alnB.fa'), nReads = 100, naming = 'filenames')
alignments2long(alns)
}
\keyword{AA,}
\keyword{DNA,}
\keyword{alignment}
