% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/between_lib_distances.R
\name{between_lib_distances}
\alias{between_lib_distances}
\title{Between library distances}
\usage{
between_lib_distances(alignments, labeling)
}
\arguments{
\item{alignments}{character vector of sequences, must be same length}

\item{labeling}{character vector of library association}
}
\description{
Plot the hamming distances between alignments from two or more libraries
}
\keyword{Library,}
\keyword{comparison}
