% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_color_hue.R
\name{gg_color_hue}
\alias{gg_color_hue}
\title{Color generator function identical to ggplot coloring}
\usage{
gg_color_hue(n)
}
\arguments{
\item{n}{integer. number of colors}
}
\value{
character vector of color values
}
\description{
ggplot color function
}
\keyword{color}
