% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamming_heatmap.R
\name{hamming_heatmap}
\alias{hamming_heatmap}
\title{Heatmap of hamming distances}
\usage{
hamming_heatmap(x, nReads)
}
\arguments{
\item{x}{character vector of filenames or named character vector with sequences and labels as names}

\item{nReads}{integer. number of reads to read in and}
}
\description{
This function makes a heatmap of hamming distances
}
\examples{
aa <- c(a = 'YNGTNQ', b = 'YGNMAK')
hamming_heatmap(aa)
}
\keyword{distances}
\keyword{hamming}
