% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_alignment.R
\name{plot_alignment}
\alias{plot_alignment}
\title{Lookup properties of amino acids}
\usage{
plot_alignment(
  Alignment,
  Reference = NULL,
  RemoveConserved = T,
  SequenceWrap = NULL,
  Coloring = "AA",
  SeqRange = NULL,
  fontscale = 1,
  order_by_distance = T
)
}
\description{
This function generates a ggplot2 plot that shows an alignment of the provided aligned sequences. The options:
Alignment: a string vector containing the aligned sequences
Reference: provide a sequence that is used as a reference. It will be on top of the plot. If a reference is not provided the first sequence in the alignment is used.
RemoveConserved: remove the positions in the alignment that are all the same, i.e. conserved.
SequenceWrap: useful to show long alignments, provide number of positions you want to show per "row"
Coloring: the way the letters will be colored, options are: 'AA', 'DNA', 'SideChainClass', 'SideChainPolarity', 'SideChainCharge', 'HydropathyIndex', 'MolecularWeight', 'Occurance'
SeqRange: provide the positions that you want to be plotted
fontscale: change the size of the letters
order_by_distance: if true the sequences will be ordered by hamming distance to the reference, if false will provide the sequences according to the order in the alignment vector
}
\examples{
aln <- c('ACGTACGT','AAAATTTT','CCCCGGGG')
names(aln) <- c('seq1','seq2','seq3')
plot_alignment(aln)
}
\keyword{AA,}
\keyword{acid,}
\keyword{amino}
\keyword{properties}
