% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_frequencies.R
\name{plot_frequencies}
\alias{plot_frequencies}
\title{Plot Frequencies of libraries}
\usage{
plot_frequencies(alignment, reference)
}
\arguments{
\item{alignment}{character vector with same length sequences to aligne}

\item{reference}{character vector with one sequence acting as reference to compare the AA to}
}
\description{
Plot the positional frequencies of amino acids in sequence alignments and color it according to match with reference
}
\keyword{DNA,}
\keyword{complement}
\keyword{reverse,}
