% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_alignments.R
\name{read_alignments}
\alias{read_alignments}
\title{Read in sequence alignments}
\usage{
read_alignments(
  input,
  nReads = -1L,
  naming = "headers",
  fileformat = "singleline"
)
}
\arguments{
\item{input}{filename of fasta without linebreaks in sequence}

\item{nReads}{integer value of number of reads to read in}

\item{naming}{'headers', 'filenames' or 'filenames-header'. Defines where the names come from.}

\item{fileformat}{'singleline', 'multiline'. Defines if the fasta file has the sequence entries in single lines or in multiple lines. singleline is faster but will give an error when dealing with multiline fastas.}
}
\value{
named character vector
}
\description{
This function reads in alignments from fasta files. The sequences need to be aligned, therefore need all have the same length. The function is capable of reading several files, but the contents need to match in string length. The output is a named vector.
}
\keyword{AA,}
\keyword{DNA,}
\keyword{alignment,}
\keyword{read}
