% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_complement.R
\name{reverse_complement}
\alias{reverse_complement}
\title{Make the reverse complement of strings of DNA}
\usage{
reverse_complement(x, reverse = T, collapsed = T)
}
\arguments{
\item{x}{character vector containing sequences with ACGTN-}

\item{reverse}{logical}

\item{collapsed}{logical. collapse sequence into string or output as seperate letters}
}
\value{
character vector
}
\description{
This function takes a vector of strings with the letters A C G T and converts them to the opposing letter, T G C A. Letters - and N are also accepted but will stay the same. Other letters are not accepted.
}
\examples{
dna <- c('ACGT', 'CGCGAAA')
reverse_complement(dna)
}
\keyword{DNA,}
\keyword{complement}
\keyword{reverse,}
