% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/within_lib_distances.R
\name{within_lib_distances}
\alias{within_lib_distances}
\title{Within library distances}
\usage{
within_lib_distances(alignments, labeling)
}
\arguments{
\item{alignments}{character vector of sequences, must be same length}

\item{labeling}{character vector of library association}
}
\description{
Plot the all vs all hamming distances of alignments from one or more libraries
}
\keyword{Library,}
\keyword{comparison}
