% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_fasta.R
\name{write_fasta}
\alias{write_fasta}
\title{Write sequences as fasta file}
\usage{
write_fasta(Sequences, Headers, File)
}
\description{
This function takes a vector of strings, the names for the Headers and the filename to write a fasta file
}
\examples{
write_fasta(Sequences = c('ACAAACCAGG','AAACCAGA'), Headers = c('Seq1','Seq2'), File = 'sequences.fasta')
}
\keyword{fasta,}
\keyword{sequences,}
\keyword{write}
