% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_properties.R
\name{aa_properties}
\alias{aa_properties}
\title{Lookup properties of amino acids}
\usage{
aa_properties(AA, property)
}
\description{
This function takes the single letter annotation of amino acids and converts it into the property of choice. Properties available: 'SideChainClass', 'SideChainPolarity', 'SideChainCharge', 'HydropathyIndex', 'MolecularWeight', 'Occurance'.
}
\examples{
aa_properties(AA = c('A','A','R','Q','Y'), property = 'HydropathyIndex')
}
\keyword{AA,}
\keyword{acid,}
\keyword{amino}
\keyword{properties}
