% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignments2long.R
\name{alignments2long}
\alias{alignments2long}
\title{Convert alignment vector to long table}
\usage{
alignments2long(input, seqrange = NULL, get_frequencies = F)
}
\arguments{
\item{input}{named character vector}

\item{seqrange}{integer vector describing the positions in the sequences to use}

\item{get_frequencies}{logical. calculat frequencies or not}
}
\value{
tibble with colnames c('Sample','Pos','AA') or c('Sample','Pos','AA','Freq')
}
\description{
This function takes a named vector of alignment sequences and converts it to a long data.frame with one letter per row. Columns indicate sequence ID and position in sequence. With get_frequencies the output can be condensed to frequency occurances of the letters per position.
}
\examples{
alignments2long(c(a = 'AAACCCGT', b = 'AAAGGGCT', c = 'AGAGAGAG'))
}
\keyword{AA,}
\keyword{DNA,}
\keyword{alignment}
