% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library_distances.R
\name{between_lib_distances}
\alias{between_lib_distances}
\title{Library distances}
\usage{
between_lib_distances(files, nReads)
}
\arguments{
\item{files}{character vector of filenames}

\item{nReads}{integer. number of reads to read in and}
}
\description{
This function takes a vector of strings with the letters A C G T and converts them to the opposing letter, T G C A. Letters - and N are also accepted but will stay the same. Other letters are not accepted.
}
\keyword{DNA,}
\keyword{complement}
\keyword{reverse,}
