% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_consensus.R
\name{generate_consensus}
\alias{generate_consensus}
\title{Generate consensus from alignments}
\usage{
generate_consensus(input, outformat = "stringvector")
}
\arguments{
\item{input}{file, character vector or alignments2long output}

\item{outformat}{'stringvector' or 'long'}
}
\description{
This function takes either Filepaths, named vector alignments, or a data.frame from alignments2long. The output is the consensus sequences in either 'long' or 'stringvector' format.
}
\examples{
generate_consensus(input = c(a = 'AAACAAA', a = 'AAACAAA', a = 'AAAAAAA'))
}
\keyword{AA,}
\keyword{DNA,}
\keyword{alignment,}
\keyword{consensus}
