% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_color_hue.R
\name{gg_color_hue}
\alias{gg_color_hue}
\title{Color generator function identical to ggplot coloring}
\usage{
gg_color_hue(n)
}
\description{
Provide the number of colors you want. The output will be a string of color values
}
\keyword{color}
