% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_alignment.R
\name{plot_alignment}
\alias{plot_alignment}
\title{Lookup properties of amino acids}
\usage{
plot_alignment(
  Alignment,
  Reference = NULL,
  RemoveConserved = T,
  SequenceWrap = NULL,
  Coloring = "AA",
  SeqRange = NULL,
  fontscale = 1,
  order_by_distance = T
)
}
\arguments{
\item{Alignment}{named character vector of same length sequences}

\item{Reference}{named character vector of one sequence}

\item{RemoveConserved}{logical. removes positions that are all the same in the alignment}

\item{SequenceWrap}{integer value of positions to show per line}

\item{Coloring}{'AA', 'DNA', 'SideChainClass', 'SideChainPolarity', 'SideChainCharge', 'HydropathyIndex', 'MolecularWeight', or 'Occurance'. Defines how the letters are colored.}

\item{SeqRange}{integers, start and stop position of sequences to use for alignments.}

\item{fontscale}{numeric value indicating the size of the letters}

\item{order_by_distance}{logical, sort sequences by hamming distance}
}
\value{
ggplot
}
\description{
This function generates a ggplot2 plot that shows an alignment of the provided aligned sequences.
}
\examples{
aln <- c('ACGTACGT','AAAATTTT','CCCCGGGG')
names(aln) <- c('seq1','seq2','seq3')
plot_alignment(aln)
}
\keyword{AA,}
\keyword{acid,}
\keyword{amino}
\keyword{properties}
