% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_alignments.R
\name{read_alignments}
\alias{read_alignments}
\title{Read in sequence alignments}
\usage{
read_alignments(input, nReads = -1L, naming = "headers")
}
\description{
This function reads in alignments from fasta files. The sequences need to be aligned, therefore need all have the same length. The function is capable of reading several files, but the contents need to match in string length. The output is a named vector.
}
\examples{
read_alignments(c('alnA.fa','alnB.fa'), nReads = 100, naming = 'filenames')
}
\keyword{AA,}
\keyword{DNA,}
\keyword{alignment,}
\keyword{read}
