% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamming_heatmap.R, R/reverse_complement.R
\name{reverse_complement}
\alias{reverse_complement}
\title{Heatmap of hamming distances}
\usage{
reverse_complement(x, reverse = T, collapsed = T)

reverse_complement(x, reverse = T, collapsed = T)
}
\description{
This function makes a heatmap of hamming distances

This function takes a vector of strings with the letters A C G T and converts them to the opposing letter, T G C A. Letters - and N are also accepted but will stay the same. Other letters are not accepted.
}
\examples{
dna <- c('ACGT', 'CGCGAAA')
reverse_complement(dna)
}
\keyword{DNA,}
\keyword{complement}
\keyword{distances}
\keyword{hamming}
\keyword{reverse,}
