% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_fasta.R
\name{write_fasta}
\alias{write_fasta}
\title{Write sequences as fasta file}
\usage{
write_fasta(Sequences, Headers, File)
}
\arguments{
\item{Sequences}{character vector containg the sequences to write into the file}

\item{Headers}{character vector containg the header names to write into the file, same order as sequences}

\item{File}{Filename to write to}
}
\description{
This function takes a vector of strings, the names for the Headers and the filename to write a fasta file
}
\keyword{fasta,}
\keyword{sequences,}
\keyword{write}
